# swagger.yaml
openapi: 3.0.3
info:
  title: ServiceRoute Integration Module API
  description: API для системы управления заявками на обслуживание с интеграцией 1С
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
  - url: https://api.example.com
    description: Продакшн сервер
...
paths:
...
  /requests:
    get:
      tags:
        - Заявки
      summary: Список всех заявок
      description: Возвращает HTML страницу со списком всех заявок
      security:
        - bearerAuth: []
      responses:
        '200':
          description: HTML страница со списком заявок
          content:
            text/html:
              schema:
                type: string
...
  /requests/create:
    post:
      tags:
        - Заявки
      summary: Создание новой заявки
      description: Создает новую заявку на обслуживание
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - customerId
                - address
              properties:
                customerId:
                  type: string
                  description: ID клиента или номер телефона
                  example: "1234567890"
                address:
                  type: string
                  description: Адрес выполнения работ
                  example: "г. Москва, ул. Ленина, д. 1"
                equipmentTypeId:
                  type: integer
                  description: ID типа оборудования (опционально)
                  example: 1
                customEquipmentType:
                  type: string
                  description: Пользовательский тип оборудования (если выбран тип "Другое")
                  example: "Принтер HP"
                problemDescription:
                  type: string
                  description: Описание проблемы клиента
                  example: "Не включается"
      responses:
        '302':
          description: Редирект на страницу созданной заявки
          headers:
            Location:
              schema:
                type: string
                example: "/requests/123"
        '400':
          description: Ошибка валидации данных
        '500':
          description: Внутренняя ошибка сервера
...
  /requests/{id}/edit:
    get:
      tags:
        - Заявки
      summary: Форма редактирования заявки
      description: Возвращает HTML форму для редактирования заявки (только для администраторов)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: HTML форма редактирования заявки
          content:
            text/html:
              schema:
                type: string
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
        '404':
          description: Заявка не найдена
  /requests/{id}/update:
    post:
      tags:
        - Заявки
      summary: Обновление заявки
      description: Обновляет информацию о заявке (только для администраторов)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - customerId
                - address
                - status
              properties:
                customerId:
                  type: string
                  description: ID клиента или номер телефона
                  example: "1234567890"
                address:
                  type: string
                  description: Адрес выполнения работ
                  example: "г. Москва, ул. Ленина, д. 1"
                status:
                  type: string
                  description: Статус заявки (new, assigned, in_progress, completed, canceled)
                  enum: [new, assigned, in_progress, completed, canceled]
                  example: "in_progress"
                engineerId:
                  type: integer
                  description: ID инженера (опционально)
                  example: 5
                equipmentTypeId:
                  type: integer
                  description: ID типа оборудования (опционально)
                  example: 1
                customEquipmentType:
                  type: string
                  description: Пользовательский тип оборудования
                  example: "Принтер HP"
                problemDescription:
                  type: string
                  description: Описание проблемы клиента
                  example: "Не включается"
      responses:
        '302':
          description: Редирект на страницу заявки
          headers:
            Location:
              schema:
                type: string
                example: "/requests/1"
        '400':
          description: Ошибка валидации данных
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
        '500':
          description: Внутренняя ошибка сервера
  /requests/{id}/assign:
    post:
      tags:
        - Заявки
      summary: Назначение заявки инженеру
      description: Назначает заявку инженеру (переводит статус из new в assigned)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - engineerId
              properties:
                engineerId:
                  type: integer
                  description: ID инженера
                  example: 5
      responses:
        '302':
          description: Редирект на страницу заявки
          headers:
            Location:
              schema:
                type: string
                example: "/requests/1"
        '400':
          description: Ошибка валидации данных
        '403':
          description: Доступ запрещен (требуется роль ADMIN или MANAGER)
        '500':
          description: Внутренняя ошибка сервера
  /requests/{id}/start:
    post:
      tags:
        - Заявки
      summary: Начало работы над заявкой
      description: Переводит заявку в статус "В работе" (assigned -> in_progress)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '302':
          description: Редирект на страницу заявки
          headers:
            Location:
              schema:
                type: string
                example: "/requests/1"
        '400':
          description: Ошибка валидации данных
        '403':
          description: Доступ запрещен (требуется роль ENGINEER)
        '500':
          description: Внутренняя ошибка сервера
  /requests/{id}/reserve-parts:
    get:
      tags:
        - Заявки
      summary: Форма резервирования запчастей
      description: Возвращает HTML форму для резервирования запчастей для заявки
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: HTML форма резервирования запчастей
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Заявка должна быть в статусе 'Назначена' или 'В работе'
        '403':
          description: Доступ запрещен (требуется роль ENGINEER)
        '404':
          description: Заявка не найдена

    post:
      tags:
        - Заявки
      summary: Резервирование запчастей для заявки
      description: Резервирует запчасти для заявки
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                partIds:
                  type: array
                  items:
                    type: integer
                  description: Список ID запчастей для резервирования
                  example: [1, 2, 3]
                quantities:
                  type: array
                  items:
                    type: integer
                  description: Количество каждой запчасти
                  example: [2, 1, 5]
      responses:
        '302':
          description: Редирект на страницу заявки
          headers:
            Location:
              schema:
                type: string
                example: "/requests/1"
        '400':
          description: Ошибка валидации данных
        '403':
          description: Доступ запрещен (требуется роль ENGINEER)
        '500':
          description: Внутренняя ошибка сервера
...
  /requests/{id}/cancel:
    post:
      tags:
        - Заявки
      summary: Отмена заявки
      description: Отменяет заявку (переводит в статус canceled)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '302':
          description: Редирект на страницу заявки
          headers:
            Location:
              schema:
                type: string
                example: "/requests/1"
        '400':
          description: Ошибка валидации данных
        '403':
          description: Доступ запрещен (требуется роль ADMIN или MANAGER)
        '500':
          description: Внутренняя ошибка сервера
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
...
security:
  - bearerAuth: []